made cif_parser and restructured parsers
