+++ molgeom/__init__.py
@@ -1,7 +1,12 @@
 from .data import consts
-from .atom import Atom as Atom
-from .molecule import Molecule as Molecule
-from .parsers import xyz_parser, gau_parser, inp_parser, poscar_parser, parse_file
+from .atom import Atom
+from .molecule import Molecule
+from .parsers.xyz import xyz_parser
+from .parsers.gaussian import gau_inp_parser
+from .parsers.gms import gms_inp_parser
+from .parsers.cif import cif_parser
+from .parsers.poscar import poscar_parser
+from .parsers.parser_selector import read_file
 from .utils.fancy_indexing_list import FancyIndexingList
 from .utils import decorators
 
@@ -10,10 +15,11 @@ __all__ = [
     "Molecule",
     "consts",
     "xyz_parser",
-    "gau_parser",
-    "inp_parser",
+    "gau_inp_parser",
+    "gms_inp_parser",
+    "cif_parser",
     "poscar_parser",
-    "parse_file",
+    "read_file",
     "FancyIndexingList",
     "decorators",
 ]
